+++ src/lib/base/hashtable.c
@@ -25,7 +25,7 @@ Hashtable* hsh_new(int est_capacity) {
   Hashtable* ht;
   int i;
   ht = (Hashtable*)smalloc(sizeof(Hashtable));
-  ht->nbuckets = ceil(est_capacity*1.0/LOADING_FACTOR);
+  ht->nbuckets = (int)ceil(est_capacity*1.0/LOADING_FACTOR);
   if (ht->nbuckets < 10) ht->nbuckets = 10;
   ht->keys = (List**)smalloc(ht->nbuckets*sizeof(List*));
   ht->vals = (List**)smalloc(ht->nbuckets*sizeof(List*));
